refactor: 테스트와 프로덕션 코드 분리
