+++ src/test/java/racingcar/model/CarsTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.controller.NumericGenerator;
-import racingcar.model.Cars;
 
 import java.util.ArrayList;
 import java.util.List;
